Move selectable-list stories to shared
